Add files via upload
With the exponential increase in the number of vehicles and world population day by
day, vehicle availability and usage on the road in recent years, finding a space for
parking the bike is becoming more and more difficult with resulting in the number of
conflicts such as traffic problems. This is about creating a reliable system that takes
over the task of identifying free slots in a parking area and keeping the record of
vehicles parked very systematic manner. This project lessens human effort at the
parking area to a great extent such as in case of searching of free slots by the driver
and calculating the payment for each vehicle using parking area. The various steps
involved in this operation are vehicle identification using RFID tags, free slot detection
using IR sensors and payment calculation is done on the basis of period of parking and
this is done with the help of real time clock.